prepare ui for adding new accounts

+++ app/src/main/java/cz/skala/trezorwallet/data/entity/Account.kt
@@ -17,7 +17,7 @@ class Account(
         val index: Int,
         val legacy: Boolean,
         val label: String?,
-        val balance: Int
+        val balance: Double
 ) {
     fun getDisplayLabel(resources: Resources): String {
         return when {
